+++ gprofiler/profilers/python.py
@@ -14,11 +14,10 @@ from psutil import Process
 
 from gprofiler.exceptions import CalledProcessError, ProcessStoppedException, StopEventSetException
 from gprofiler.log import get_logger_adapter
-from gprofiler.main import positive_integer
 from gprofiler.merge import parse_and_remove_one_collapsed, parse_many_collapsed
 from gprofiler.profilers.profiler_base import ProcessProfilerBase, ProfilerBase, ProfilerInterface
 from gprofiler.profilers.registry import ProfilerArgument, register_profiler
-from gprofiler.types import ProcessToStackSampleCounters, StackToSampleCount
+from gprofiler.types import ProcessToStackSampleCounters, StackToSampleCount, positive_integer
 from gprofiler.utils import (
     pgrep_maps,
     poll_process,
